Move logic to base controller
